move default fonts to theme package, better folder structure
