+++ packages/fluent-theme/src/fluent/FluentColors.ts
@@ -1,1 +1,1 @@
-export * from '@fluentui/theme/lib/constants/FluentColors';
+export * from '@fluentui/theme/lib/colors/FluentColors';
