+++ packages/fluent-theme/src/fluent/FluentMotion.ts
@@ -1,1 +1,1 @@
-export * from '@fluentui/theme/lib/constants/FluentMotions';
+export * from '@fluentui/theme/lib/motions/FluentMotions';
