+++ packages/theme/src/colors/DefaultPalette.ts
