+++ packages/theme/src/colors/index.ts
@@ -1,0 +1,2 @@
+export * from './FluentColors';
+export { DefaultPalette } from './DefaultPalette';
