--- packages/theme/src/constants/index.ts
@@ -1,6 +1,0 @@
-export * from './FluentColors';
-export * from './FluentFonts';
-export * from './FluentMotions';
-export { DefaultPalette } from './DefaultPalette';
-export { DefaultEffects } from './DefaultEffects';
-export { DefaultSpacing } from './DefaultSpacing';
