+++ packages/theme/src/effects/index.ts
@@ -1,0 +1,1 @@
+export { DefaultEffects } from './DefaultEffects';
