+++ packages/theme/src/fonts/index.ts
@@ -1,0 +1,3 @@
+export * from './FluentFonts';
+export { createFontStyles } from './createFontStyles';
+export { DefaultFontStyles, registerDefaultFontFaces } from './DefaultFontStyles';
