+++ packages/theme/src/index.ts
@@ -1,3 +1,7 @@
 export * from './mergeThemes';
 export * from './types/index';
-export * from './constants/index';
+export * from './colors/index';
+export * from './effects/index';
+export * from './spacing/index';
+export * from './motions/index';
+export * from './fonts/index';
