+++ packages/theme/src/spacing/index.ts
@@ -1,0 +1,1 @@
+export { DefaultSpacing } from './DefaultSpacing';
