+++ packages/theme/src/types/IFabricConfig.ts
@@ -1,4 +1,8 @@
+/* eslint-disable @typescript-eslint/naming-convention */
 import { IStyleSheetConfig } from '@uifabric/merge-styles';
+
+// TODO: figure out a better place (e.g a shared typing package) for this interface.
+
 /**
  * The interface of window.FabricConfig, which can be burned on the page before script loading to preemptively
  * define default configurations.
